Finished backend with pydantic schema. Tried some changes on grid/table for features in UI, but after constant issues, will revert to similar experience to Datasets table.
